Реализация функции подсчета стандартных выплат по страховой пенсии по годам

+++ calculator-backend/src/utils/payments/types/paymentType.py
@@ -1,6 +1,6 @@
 from typing import  Dict, TypeAlias
 
-AmountByYear: TypeAlias = Dict[int, int]
+AmountByYear: TypeAlias = Dict[int, float]
 
-PaymentsByYear: TypeAlias = Dict[Dict[int, int]]
+PaymentsByYear: TypeAlias = Dict[Dict[int, float]]
